requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-openai
langchain-core
pypdf
faiss-cpu
tiktoken
openai
